Move to top level class
