+++ spring/boot-autoconfigure/src/test/java/com/linecorp/armeria/spring/ArmeriaAutoConfigurationTest.java
@@ -72,7 +72,6 @@ import com.linecorp.armeria.server.grpc.GrpcServiceBuilder;
 import com.linecorp.armeria.server.logging.LoggingService;
 import com.linecorp.armeria.server.thrift.THttpService;
 import com.linecorp.armeria.spring.ArmeriaAutoConfigurationTest.TestConfiguration;
-import com.linecorp.armeria.spring.GrpcServiceRegistrationBean.GrpcExampleRequest;
 import com.linecorp.armeria.spring.test.grpc.main.Hello.HelloReply;
 import com.linecorp.armeria.spring.test.grpc.main.Hello.HelloRequest;
 import com.linecorp.armeria.spring.test.grpc.main.HelloServiceGrpc;
